Refactoring of paths due to changes of file structure
